Add outbound api hook refactor
