+++ packages/invoices/src/OutboundShipment/DetailView/DetailView.tsx
@@ -17,15 +17,15 @@ import { Toolbar } from './Toolbar';
 import { Footer } from './Footer';
 import { AppBarButtons } from './AppBarButtons';
 import { SidePanel } from './SidePanel';
-import { useOutbound, useIsOutboundDisabled } from '../api';
+import { useOutbound } from '../api';
 import { AppRoute } from '@openmsupply-client/config';
 import { OutboundLineFragment } from '../api/operations.generated';
 
 export const DetailView: FC = () => {
-  const isDisabled = useIsOutboundDisabled();
+  const isDisabled = useOutbound.utils.isDisabled();
   const { entity, mode, onOpen, onClose, isOpen } =
     useEditModal<ItemRowFragment>();
-  const { data, isLoading } = useOutbound();
+  const { data, isLoading } = useOutbound.document.get();
   const t = useTranslation('distribution');
   const navigate = useNavigate();
   const onRowClick = useCallback(
