+++ packages/invoices/src/OutboundShipment/ListView/ListView.tsx
@@ -14,7 +14,7 @@ import {
 import { getStatusTranslator, isOutboundDisabled } from '../../utils';
 import { Toolbar } from './Toolbar';
 import { AppBarButtons } from './AppBarButtons';
-import { useOutbounds, useUpdateOutbound } from '../api';
+import { useOutbound } from '../api';
 import { OutboundRowFragment } from '../api/operations.generated';
 
 const useDisableOutboundRows = (rows?: OutboundRowFragment[]) => {
@@ -26,7 +26,7 @@ const useDisableOutboundRows = (rows?: OutboundRowFragment[]) => {
 };
 
 export const OutboundShipmentListViewComponent: FC = () => {
-  const { mutate: onUpdate } = useUpdateOutbound();
+  const { mutate: onUpdate } = useOutbound.document.update();
   const t = useTranslation('common');
   const navigate = useNavigate();
 
@@ -38,7 +38,7 @@ export const OutboundShipmentListViewComponent: FC = () => {
     onChangePage,
     pagination,
     filter,
-  } = useOutbounds();
+  } = useOutbound.document.list();
   useDisableOutboundRows(data?.nodes);
 
   const { c } = useCurrency();
