+++ packages/invoices/src/OutboundShipment/api/hooks/document/useOutboundDelete.ts
@@ -1,0 +1,53 @@
+import { OutboundRowFragment } from './../../operations.generated';
+import { useOutboundApi } from './../utils/useOutboundApi';
+import {
+  useQueryClient,
+  useTranslation,
+  useMutation,
+  useNotification,
+  useTableStore,
+} from '@openmsupply-client/common';
+import { useOutbounds } from './useOutbounds';
+import { canDeleteInvoice } from '../../../../utils';
+
+export const useOutboundDelete = () => {
+  const queryClient = useQueryClient();
+  const { data: rows } = useOutbounds();
+  const api = useOutboundApi();
+  const { mutate } = useMutation(api.delete);
+  const t = useTranslation('replenishment');
+
+  const { success, info } = useNotification();
+
+  const { selectedRows } = useTableStore(state => ({
+    selectedRows: Object.keys(state.rowState)
+      .filter(id => state.rowState[id]?.isSelected)
+      .map(selectedId => rows?.nodes?.find(({ id }) => selectedId === id))
+      .filter(Boolean) as OutboundRowFragment[],
+  }));
+
+  const deleteAction = () => {
+    const numberSelected = selectedRows.length;
+    if (selectedRows && numberSelected > 0) {
+      const canDeleteRows = selectedRows.every(canDeleteInvoice);
+      if (!canDeleteRows) {
+        const cannotDeleteSnack = info(t('messages.cant-delete-invoices'));
+        cannotDeleteSnack();
+      } else {
+        mutate(selectedRows, {
+          onSuccess: () => queryClient.invalidateQueries(['invoice']),
+        });
+        const deletedMessage = t('messages.deleted-invoices', {
+          number: numberSelected,
+        });
+        const successSnack = success(deletedMessage);
+        successSnack();
+      }
+    } else {
+      const selectRowsSnack = info(t('messages.select-rows-to-delete'));
+      selectRowsSnack();
+    }
+  };
+
+  return deleteAction;
+};
