+++ packages/invoices/src/OutboundShipment/api/hooks/line/useOutboundSaveLines.ts
@@ -1,0 +1,14 @@
+import { useQueryClient, useMutation } from '@openmsupply-client/common';
+import { useOutboundApi } from './../utils/useOutboundApi';
+import { useOutboundNumber } from './../utils/useOutboundNumber';
+
+export const useOutboundSaveLines = () => {
+  const outboundNumber = useOutboundNumber();
+  const queryClient = useQueryClient();
+  const api = useOutboundApi();
+  return useMutation(api.updateLines, {
+    onSuccess: () => {
+      queryClient.invalidateQueries(api.keys.detail(outboundNumber));
+    },
+  });
+};
