+++ packages/invoices/src/OutboundShipment/api/hooks/utils/index.ts
@@ -1,0 +1,11 @@
+import { useOutboundNumber } from './useOutboundNumber';
+import { useOutboundIsDisabled } from './useOutboundIsDisabled';
+import { useOutboundApi } from './useOutboundApi';
+
+export { useOutboundApi, useOutboundIsDisabled, useOutboundNumber };
+
+export const utils = {
+  useOutboundNumber,
+  useOutboundIsDisabled,
+  useOutboundApi,
+};
