+++ packages/invoices/src/OutboundShipment/api/hooks/utils/useOutboundNumber.ts
@@ -1,0 +1,6 @@
+import { useParams } from '@openmsupply-client/common';
+
+export const useOutboundNumber = () => {
+  const { invoiceNumber = '' } = useParams();
+  return invoiceNumber;
+};
